requirements.txt
agno==2.5.9
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
azure-ai-inference==1.0.0b9
azure-core==1.38.3
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.5
click==8.3.1
colorama==0.4.6
cryptography==46.0.5
curl_cffi==0.13.0
ddgs==9.11.3
deprecation==2.1.0
distro==1.9.0
docstring_parser==0.17.0
duckduckgo_search==8.1.1
fastapi==0.135.1
frozendict==2.4.7
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-auth==2.49.1
google-genai==1.67.0
greenlet==3.3.2
groq==1.1.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
isodate==0.7.2
jiter==0.13.0
lance-namespace==0.5.2
lance-namespace-urllib3-client==0.5.2
lancedb==0.29.2
lxml==6.0.2
markdown-it-py==4.0.0
mdurl==0.1.2
multidict==6.7.1
multitasking==0.0.12
numpy==2.4.3
openai==2.26.0
packaging==26.0
pandas==3.0.1
peewee==4.0.1
platformdirs==4.9.4
primp==1.1.3
propcache==0.4.1
protobuf==7.34.0
pyarrow==23.0.1
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
pylance==3.0.0
pypdf==6.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pytz==2026.1.post1
PyYAML==6.0.3
requests==2.32.5
rich==14.3.3
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
soupsieve==2.8.3
SQLAlchemy==2.0.48
sseclient-py==1.9.0
starlette==0.52.1
tantivy==0.25.1
tenacity==9.1.4
tqdm==4.67.3
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.42.0
websockets==16.0
yarl==1.23.0
yfinance==1.2.0